REGOD-BE/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.34
psycopg2-binary==2.9.10
python-dotenv==1.0.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
requests==2.31.0
python-multipart==0.0.6
alembic==1.12.1
docker==6.1.2